Introduce CrayonsHelper
